DINOv2/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
pydicom>=2.4.0
pandas
dinov2
timm
xformers --index-url https://download.pytorch.org/whl/cu121
globus-sdk>=3.0.0
mlflow
scikit-learn
seaborn
matplotlib
tqdm
pyyaml
